fixing major code duplications in expandvar and preprocess
